allow Spawner to be automatically placed as well
- can not use multiple custom editors for the same type
