+++ Assets/Scripts/EditorHelpers/PlacementHelperEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 44ce3bf120e3c6f45b048110c9361855
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
